Move Dashboard component into a dedicated folder and pull its props and state definition out of the dashboard component file. Use it accordingly in the Root component.
