+++ app/components/screens/dashboard/Props.ts
@@ -1,0 +1,13 @@
+export interface OwnProps {
+
+}
+
+export interface StateProps {
+  //no-op
+}
+
+export interface DispatchProps {
+  //no-op
+}
+
+export type Props = OwnProps & StateProps & DispatchProps
