+++ app/components/screens/dashboard/State.ts
@@ -1,0 +1,8 @@
+export enum NavigationState {
+  Here,
+  ProductsScreen
+}
+
+export interface State {
+  navigationState: NavigationState
+}
